Rename names form RestaurantProfile to EatingPlaceProfile in director EatingPlaceProfile
