+++ src/components/OwnerEatingPlace/EatingPlaceProfile/EatingPlaceProfile.js
@@ -1,0 +1,14 @@
+import React from "react";
+import EatingPlaceHeader from "./EatingPlaceHeader/EatingPlaceHeader";
+import EatingPlaceBody from "./EatingPlaceBody/RestaurantBody";
+
+const EatingPlaceProfile = props => {
+  return (
+    <div>
+      <EatingPlaceHeader />
+      <EatingPlaceBody />
+    </div>
+  );
+};
+
+export default EatingPlaceProfile;
