--- src/components/OwnerEatingPlace/EatingPlaceProfile/RestaurantProfile.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import RestaurantHeader from './Header/Header';
-import RestaurantBody from './RestaurantBody/RestaurantBody';
-
-const RestaurantProfile = props => {
-   return (
-       <div>
-           <RestaurantHeader/>
-            <RestaurantBody/>
-       </div>
-   )
-}
-
-export default RestaurantProfile;
